requirements.txt
asgiref==3.8.1
Django==5.1
django-auditlog==3.0.0
django-cors-headers==4.4.0
django-treebeard==4.7.1
djangorestframework==3.15.2
phonenumbers==8.13.43
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
